% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_email.R
\name{clean_processed_payment_data_from_email}
\alias{clean_processed_payment_data_from_email}
\title{Clean the output of get_processed_payment_data_from_email and format for alignment
with invoice_line_item data}
\usage{
clean_processed_payment_data_from_email(processed_payment_data_from_email)
}
\arguments{
\item{processed_payment_data_from_email, }{the dataframe output of get_processed_payment_data_from_email}
}
\value{
The inpout data frame reformatted for alignment with invoice_line_item
}
\description{
Clean the output of get_processed_payment_data_from_email and format for alignment
with invoice_line_item data
}
\examples{
\dontrun{
processed_payment_data <- clean_processed_payment_data_from_email(
  processed_payment_data_from_email
)
}
}
