% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clean_processed_payment_data_from_email_test_data}
\alias{clean_processed_payment_data_from_email_test_data}
\title{clean_processed_payment_data_from_email_test_data}
\format{
A named list of 2 dataframes:
\describe{
  \item{\code{input}}{a dataframe of test data as returned from get_processed_payment_data_from_email}
  \item{\code{output}}{a dataframe of test data corresponding to the input data frame}
}
}
\usage{
clean_processed_payment_data_from_email_test_data
}
\description{
A named list of dataframes used to test the function clean_processed_payment_data_from_email_test_data
}
\details{
DETAILS
}
\keyword{datasets}
