% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cleanup_project_ownership_test_data}
\alias{cleanup_project_ownership_test_data}
\title{cleanup_project_ownership_test_data}
\format{
A named list of 5 dataframes:
\describe{
  \item{\code{redcap_user_information}}{REDCap Core table}
  \item{\code{redcap_projects}}{REDCap Core table}
  \item{\code{redcap_user_rights}}{REDCap Core table}
  \item{\code{redcap_user_roles}}{REDCap Core table}
  \item{\code{redcap_project_last_users}}{local table}
}
}
\usage{
cleanup_project_ownership_test_data
}
\description{
A named list of dataframes used to test the functions written for cleanup_project_ownership.R
}
\details{
DETAILS
}
\keyword{datasets}
