% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcc.billing.R
\name{connect_to_rcc_billing_db}
\alias{connect_to_rcc_billing_db}
\title{Connect to the rcc_billing DB}
\usage{
connect_to_rcc_billing_db(drv = RMariaDB::MariaDB(), continue_on_error = FALSE)
}
\arguments{
\item{drv, }{an object that inherits from DBIDriver (e.g. RMariaDB::MariaDB()), or an existing DBIConnection object (in order to clone an existing connection).}

\item{continue_on_error}{if TRUE then continue execution on error, if FALSE then quit non interactive sessions on error}
}
\value{
An S4 object. Run ?dbConnect for more information
}
\description{
Connect to the rcc_billing DB
}
\examples{

\dontrun{
# connect to the RCC Billing database using RCCBILLING_* environment variables
con <- connect_to_rcc_billing_db()

# connect to sqlite RCC Billing db
con <- connect_to_rcc_billing_db(drv = RSQLite::SQLite())
}
}
