% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{convert_schema_to_sqlite}
\alias{convert_schema_to_sqlite}
\title{Converts a mysql schema to a sqlite schema for the provided table.
A corresponding schema file with the same name is required in /schema.}
\usage{
convert_schema_to_sqlite(table_name)
}
\arguments{
\item{table_name, }{the name of the table to convert}
}
\description{
Converts a mysql schema to a sqlite schema for the provided table.
A corresponding schema file with the same name is required in /schema.
}
\examples{
\dontrun{
convert_schema_to_sqlite(table_name = "service_type")
}
}
