% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{create_table}
\alias{create_table}
\title{Creates a table for table_name.
A corresponding data set and sqlite schema are required in /data and /schema respectively.
Use \code{\link{convert_schema_to_sqlite}} to generate a sqlite schema from a mysql schema.}
\usage{
create_table(conn, sqlite_schema)
}
\arguments{
\item{conn, }{a DBI connection object}

\item{sqlite_schema, }{the ddl to execute against conn}
}
\description{
Creates a table for table_name.
A corresponding data set and sqlite schema are required in /data and /schema respectively.
Use \code{\link{convert_schema_to_sqlite}} to generate a sqlite schema from a mysql schema.
}
\examples{
\dontrun{
 table_name <- "service_type"
 conn <- DBI::dbConnect(RSQLite::SQLite(), dbname = ":memory:")

 schema <- convert_schema_to_sqlite(table_name)
 create_table(conn = conn, sqlite_schema = schema)
}
}
