% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csbt_column_names}
\alias{csbt_column_names}
\title{csbt_column_names}
\format{
A data frame with 12 rows and 2 variables:
\describe{
  \item{\code{ctsit}}{character: Internal database columns}
  \item{\code{csbt}}{character: Billing invoice database column names}
}
}
\usage{
csbt_column_names
}
\description{
Column names useful for CTSI-IT billing
}
\keyword{datasets}
