% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctsit_staff}
\alias{ctsit_staff}
\title{CTS-IT Staff}
\format{
A data frame with 6 rows and 1 variable:
\describe{
  \item{\code{redcap_username}}{character: a REDCap username. Typically this is Gatorlink ID.}
}
}
\usage{
ctsit_staff
}
\description{
usernames of CTS-IT staff
}
\keyword{datasets}
