% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctsit_staff_employment_periods}
\alias{ctsit_staff_employment_periods}
\title{ctsit_staff_employment_periods}
\format{
A data frame with 7 rows and 2 variables:
\describe{
  \item{\code{redcap_username}}{character a REDCap username. Typically this is Gatorlink ID.}
  \item{\code{employment_interval}}{double a lubridate time interval}
}
}
\usage{
ctsit_staff_employment_periods
}
\description{
Usernames and employment periods of CTS-IT staff. This
dataset will be used to inform default data ownership and setting billable flags
in the REDCap Entity / Project Ownership table. Multiple employment periods are
represented as multiple rows
}
\keyword{datasets}
