% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{fix_data_in_redcap_log_event}
\alias{fix_data_in_redcap_log_event}
\title{fix_data_in_redcap_log_event}
\usage{
fix_data_in_redcap_log_event(data)
}
\arguments{
\item{data}{- a dataframe containing data from the redcap_log_event tables}
}
\value{
The input dataframe with revised data types
}
\description{
Fixes column data types that can vary between MySQL/MariaDB and SQLite3.
This allows testing in SQLite3 while production is MariaDB
}
\examples{
\dontrun{
fix_data_in_redcap_log_event(redcap_log_event_test_data)
}
}
