% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{fix_data_in_redcap_user_information}
\alias{fix_data_in_redcap_user_information}
\title{fix_data_in_redcap_user_information}
\usage{
fix_data_in_redcap_user_information(data)
}
\arguments{
\item{data}{- a dataframe}
}
\value{
The input dataframe with revised data types
}
\description{
Fixes column data types that can vary between MySQL/MariaDB and SQLite3. This allows testing in SQLite while production is MariaDB
}
\examples{
\dontrun{
invoice_line_item_df_from(invoice_line_item_communications_test_data)
}
}
