% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{generate_fiscal_year_interval}
\alias{generate_fiscal_year_interval}
\title{Generate a fiscal year interval}
\usage{
generate_fiscal_year_interval(csbt_label)
}
\arguments{
\item{csbt_label}{A string representing the fiscal year range, in the format YYYY-YYYY}
}
\value{
A lubridate interval object from the start of the UF fiscal year (July 1) to the end
}
\description{
Generate a fiscal year interval
}
\examples{
\dontrun{
fiscal_year_interval <- generate_fiscal_year_interval("2021-2022")
}
}
