% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_scraping_utils.R
\name{get_bad_emails_from_log}
\alias{get_bad_emails_from_log}
\title{Get bad email adresses from the billing log}
\usage{
get_bad_emails_from_log()
}
\value{
A vector of email adresses that resulted in an error in the rcc_job_log table
}
\description{
Get bad email adresses from the billing log
}
\examples{

\dontrun{
bad_recipients <- get_bad_emails_from_log()
}
}
