% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_billable_candidates.R
\name{get_billable_candidates}
\alias{get_billable_candidates}
\title{get_billable_candidates}
\usage{
get_billable_candidates(rc_conn, rcc_billing_conn)
}
\arguments{
\item{rc_conn}{- DBI connection object to a REDCap database}

\item{rcc_billing_conn}{- DBI connection object to an rcc.billing database}
}
\value{
A dataframe of REDCap projects with relevant billing details
}
\description{
Return a data frame of REDCap projects with relevant billing details
}
\examples{
\dontrun{

library(tidyverse)
library(rcc.billing)
library(lubridate)
library(DBI)
library(dotenv)
library(redcapcustodian)

init_etl("billable_candidates")

rc_conn <- connect_to_redcap_db()
rcc_billing_conn <- connect_to_rcc_billing_db()

billable_candidates <- get_billable_candidates(rc_conn, rcc_billing_conn)
}
}
