% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_projects.R
\name{get_creators}
\alias{get_creators}
\title{get_creators}
\usage{
get_creators(
  redcap_projects,
  redcap_user_information,
  redcap_staff_employment_periods,
  include_suspended_users = FALSE,
  return_project_ownership_format = FALSE
)
}
\arguments{
\item{redcap_projects, }{The contents of the REDCap table of the same name.}

\item{redcap_user_information, }{The contents of the REDCap table of the same name.}

\item{redcap_staff_employment_periods, }{a dataset of redcap usernames and employment
intervals with one interval per row}

\item{include_suspended_users, }{Include users whose accounts are suspended}

\item{return_project_ownership_format, }{Rename the columns to match the
redcap_entity_project_ownership format}
}
\value{
a dataframe of project creators from redcap_projects
}
\description{
Returns a dataframe of project creator usernames for non-suspended,
 non-redcap-staff, with a primary email address. Suspended creators
 can be optionally included.
}
\examples{
\dontrun{
unsuspended_creators <- get_creators(
  redcap_projects = redcap_projects,
  redcap_user_information = redcap_user_information,
  redcap_staff_employment_periods = ctsit_staff_employment_periods,
  return_project_ownership_format = T
)

creators <- get_creators(
  redcap_projects = redcap_projects,
  redcap_user_information = redcap_user_information,
  redcap_staff_employment_periods = ctsit_staff_employment_periods,
  include_suspended_users = T,
  return_project_ownership_format = T
)
}
}
