% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctsi_study_id_to_project_id_map.R
\name{get_ctsi_study_id_to_project_id_map}
\alias{get_ctsi_study_id_to_project_id_map}
\title{Map CTSI Study IDs to Project IDs}
\usage{
get_ctsi_study_id_to_project_id_map(service_requests, rcc_billing_conn)
}
\arguments{
\item{service_requests}{A data frame of service requests, REDCap Service Request PID 1414.}

\item{rcc_billing_conn}{A connection to REDCap billing database. \code{\link{connect_to_rcc_billing_db}}}
}
\value{
A data frame with distinct project_id and ctsi_study_id columns, representing the mapping between project IDs and CTSI study IDs.
}
\description{
This function generates a mapping between CTSI study IDs and project IDs from invoice_line_item and REDCap Service Request project.
It filters invoice line items for service type codes 1 and 2, ensuring CTSI study IDs are present, and collects distinct pairs of project IDs and CTSI study IDs.
}
\examples{
\dontrun{
get_ctsi_study_id_to_project_id_map(service_requests, rcc_billing_conn)
}
}
