% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_fiscal_years}
\alias{get_fiscal_years}
\title{Generates a dataframe of fiscal years}
\usage{
get_fiscal_years()
}
\value{
A dataframe of fiscal years from 2019-2020 to 2039-2040
\itemize{
  \item csbt_label - A string representing the fiscal year range, in the format YYYY-YYYY
  \item fy_interval - A lubridate interval object from the start of the UF fiscal year (July 1) to the end
}
}
\description{
Generates a dataframe of fiscal years
}
\examples{
\dontrun{
fiscal_years <- get_fiscal_years()
}
}
