% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_projects.R
\name{get_last_project_user}
\alias{get_last_project_user}
\title{get_last_project_user}
\usage{
get_last_project_user(con, pid)
}
\arguments{
\item{con, }{a DBI connection object to a REDCap database}

\item{pid, }{the project ID of the project of interest.}
}
\value{
Username of the last user to log an actio against the project
}
\description{
Returns the last user to log an event on a project. This function is not vectorized.
}
\examples{
\dontrun{
  get_last_project_user(
    con = redcap_connection,
    pid = project_id
  )
}
}
