% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_project_invoice_line_items.R
\name{get_new_project_invoice_line_items}
\alias{get_new_project_invoice_line_items}
\title{Get new project billing invoice line items given a dataframe of projects to invoice,
the initial state of the invoice_line_item table,
a connection to the redcap database, and a connection to the rcc billing database.}
\usage{
get_new_project_invoice_line_items(
  projects_to_invoice,
  initial_invoice_line_item,
  rc_conn,
  rcc_billing_conn,
  api_uri
)
}
\arguments{
\item{projects_to_invoice}{a dataframe of projects to invoice}

\item{initial_invoice_line_item}{a dataframe with the initial state of the invoice_line_item table}

\item{rc_conn}{a DBI connection to the REDCap database}

\item{rcc_billing_conn}{a DBI connection to the rcc billing database}

\item{api_uri}{the URI to the redcap host's API interface}
}
\value{
a data frame of new invoice line items
}
\description{
Get new project billing invoice line items given a dataframe of projects to invoice,
the initial state of the invoice_line_item table,
a connection to the redcap database, and a connection to the rcc billing database.
}
\examples{
\dontrun{
get_new_project_invoice_line_items <- function(
  projects_to_invoice,
  initial_invoice_line_item,
  rc_conn,
  rcc_billing_conn,
  api_uri
}
}
