% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_project_service_instances.R
\name{get_new_project_service_instances}
\alias{get_new_project_service_instances}
\title{Get new service instances that need to be created given a dataframe of
projects_to_invoice and a dataframe of the initial service_instance records.}
\usage{
get_new_project_service_instances(
  projects_to_invoice,
  initial_service_instance
)
}
\arguments{
\item{projects_to_invoice}{- the projects for which we will need to create invoice line items}

\item{initial_service_instance}{- a dataframe of the existing service instance table}
}
\value{
a dataframe of new service_instance rows
}
\description{
Get new service instances that need to be created given a dataframe of
projects_to_invoice and a dataframe of the initial service_instance records.
}
\examples{
\dontrun{
  get_new_project_service_instances(projects_to_invoice, initial_service_instance)
}
}
