% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orphaned_projects.R
\name{get_orphaned_projects}
\alias{get_orphaned_projects}
\title{get_orphaned_projects}
\usage{
get_orphaned_projects(rc_conn, rcc_billing_conn, months_previous = 0)
}
\arguments{
\item{rc_conn}{- a connection to a redcap database, \code{\link[redcapcustodian]{connect_to_redcap_db}}}

\item{rcc_billing_conn}{- a connection to an rcc_billing database, \code{\link{connect_to_rcc_billing_db}}}

\item{months_previous}{- the nth month previous to today to consider}
}
\value{
a dataframe describing orphaned projects
\itemize{
  \item project_id - project_id of the orphaned project
  \item reason - why this project was selected
  \item priority - the priority of the reason
}
}
\description{
Return a dataframe of projects that have been orphaned
}
\examples{
\dontrun{
get_orphaned_projects(
  rc_conn = rc_conn,
  rcc_billing_conn = rcc_billing_conn,
  months_previous = 0
)
}
}
