% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_email.R
\name{get_processed_payment_data_from_email}
\alias{get_processed_payment_data_from_email}
\title{Scrape an inbox for processed payment emails and return data}
\usage{
get_processed_payment_data_from_email(
  username,
  password,
  url = "imaps://outlook.office365.com",
  messages_since_date
)
}
\arguments{
\item{username}{The username of the IMAP mailbox}

\item{password}{The password of the IMAP mailbox}

\item{url}{The IMAP URL of the host that houses the mailbox}

\item{messages_since_date}{The sent date of the oldest message that should be inspected}
}
\value{
A dataframe of processed payment facts
\itemize{
  \item{\code{study_name}}{character Study Name}
  \item{\code{crc_number}}{character Clinical Research Center Number}
  \item{\code{ids_number}}{character Investigative Drug Service Number}
  \item{\code{ocr_number}}{character Office of Clinical Research Number}
  \item{\code{invoice_number}}{character Invoice Number}
  \item{\code{amount_paid}}{character amount paid on an invoice}
  \item{\code{je_number}}{character Journal Entry number}
  \item{\code{je_posting_date}}{character Journal Entry Posting Date}
  \item{\code{ctsi_study_id}}{character CTSI numeric identifier for a service applied to a study}
  \item{\code{month_invoiced}}{character the month the invoice was sent, e.g., "April"}
  \item{\code{fiscal_year}}{character the fiscal year in wich the invoice was sent, e.g. "2021-2022"}
  \item{\code{sender}}{character email sender}
  \item{\code{recipient}}{character email recipient}
  \item{\code{date_sent}}{POSIX_CT date the email was sent}
}
}
\description{
Connect to an imap mailbox, identify processed payment emails sent
after `messages_since_date`, and extract the data from those emails.
}
\examples{
\dontrun{
get_processed_payment_data_from_email(
  username = "jdoe",
  password = "jane_does_password",
  url ="imaps://outlook.office365.com",
  messages_since_date = as.Date("2022-01-01", format = "\%Y-\%m-\%d")
  )
}
}
