% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_details_for_billing.R
\name{get_project_details_for_billing}
\alias{get_project_details_for_billing}
\title{Get Project Details for Billing}
\usage{
get_project_details_for_billing(
  rc_conn,
  rcc_billing_conn,
  project_ids = NA_real_
)
}
\arguments{
\item{rc_conn}{A REDCap database connection, e.g. the object returned from \code{\link[redcapcustodian]{connect_to_redcap_db}}}

\item{rcc_billing_conn}{A connection to REDCap billing database. \code{\link{connect_to_rcc_billing_db}}}

\item{project_ids}{An optional vector of project IDs to retrieve details for.}
}
\value{
A data frame with project details.
}
\description{
This function retrieves detailed information about specific projects for billing purposes. It queries:
\itemize{
 \item redcap_projects
 \item redcap_entity_project_ownership
 \item redcap_user_information
 \item invoice_line_item
 }
}
\examples{
\dontrun{
rc_conn <- connect_to_redcap_db()
rcc_billing_conn <- connect_to_rcc_billing_db()
project_ids <- c(12, 14, 22)
project_details <- get_project_details_for_billing(rc_conn, rcc_billing_con, project_ids)
project_details <- get_project_details_for_billing(rc_conn, rcc_billing_con)
}

}
