% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_projects.R
\name{get_project_pis}
\alias{get_project_pis}
\title{get_project_pis}
\usage{
get_project_pis(redcap_projects, return_project_ownership_format = FALSE)
}
\arguments{
\item{redcap_projects, }{The contents of the REDCap table of the same name.}

\item{return_project_ownership_format, }{Rename the columns to match the
redcap_entity_project_ownership format}
}
\value{
a dataframe of project_PI details from redcap_projects
}
\description{
Returns a dataframe of all project_PI details in redcap_projects for PIs with an email address in project_pi_email
}
\examples{
\dontrun{
get_project_pis(
  redcap_projects = redcap_projects,
  return_project_ownership_format = TRUE
)
}
}
