% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_projects.R
\name{get_reassigned_line_items}
\alias{get_reassigned_line_items}
\title{Get a dataframe of reassigned line items}
\usage{
get_reassigned_line_items(sent_line_items, rc_conn)
}
\arguments{
\item{sent_line_items}{a dataframe returned by \code{\link{get_unpaid_redcap_prod_per_project_line_items}}}

\item{rc_conn}{- A REDCap database connection, e.g. the object returned from \code{\link[redcapcustodian]{connect_to_redcap_db}}}
}
\value{
A dataframe of revised redcap project invoice_line_items reassigned to new owners
}
\description{
After invoices are sent, investigators will report that some projects do
not belong to them. CTS-IT will change the ownership of these projects to a
new owner. In other cases, the customer will change the ownership of a project themselves.
CTS-IT will create a report of differences in project ownership and investigators of each unpaid,
invoiced project. CTS-IT will update its tables and tell CSBT about the change.
}
\examples{
\dontrun{
rc_conn <- connect_to_redcap_db()
rcc_billing_conn <- connect_to_rcc_billing_db()

sent_line_items <- get_unpaid_redcap_prod_per_project_line_items(rcc_billing_conn)
reassigned_line_items <- get_reassigned_line_items(sent_line_items, rc_conn)
}
}
\seealso{
\code{\link{csbt_column_names}}
}
