% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_projects.R
\name{get_research_projects_not_using_viable_pi_data}
\alias{get_research_projects_not_using_viable_pi_data}
\title{get_research_projects_not_using_viable_pi_data}
\usage{
get_research_projects_not_using_viable_pi_data(
  redcap_projects,
  redcap_entity_project_ownership,
  redcap_user_information
)
}
\arguments{
\item{redcap_projects, }{The contents of the REDCap table of the same name.}

\item{redcap_entity_project_ownership, }{The contents of the REDCap Project Ownership table of the same name.}

\item{redcap_user_information, }{The contents of the REDCap table of the same name.}
}
\value{
a vector of project IDs
}
\description{
Returns the project_ids of projects that have no viable PI data
}
\examples{
\dontrun{
get_research_projects_not_using_viable_pi_data(
  redcap_projects = redcap_projects,
  redcap_entity_project_ownership = redcap_entity_project_ownership,
  redcap_user_information = redcap_user_information
)
}
}
