% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_target_projects_to_invoice.R
\name{get_target_projects_to_invoice}
\alias{get_target_projects_to_invoice}
\title{Get details on the projects we need to create invoice line items for.}
\usage{
get_target_projects_to_invoice(rc_conn)
}
\arguments{
\item{rc_conn}{A connection to a REDCap database that uses the project_ownership module}
}
\value{
a dataframe of project and owner details
}
\description{
Get details on the projects we need to create invoice line items for.
}
\examples{
\dontrun{
  get_target_projects_to_invoice(rc_conn)
}
}
