% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_projects.R
\name{get_unpaid_redcap_prod_per_project_line_items}
\alias{get_unpaid_redcap_prod_per_project_line_items}
\title{Create a dataframe of unpaid REDCap production per project line items that were sent}
\usage{
get_unpaid_redcap_prod_per_project_line_items(rcc_billing_conn)
}
\arguments{
\item{rcc_billing_conn}{- A connection to REDCap billing database. \code{\link{connect_to_rcc_billing_db}}}
}
\value{
A dataframe containing unpaid sent line items
}
\description{
Create a dataframe of unpaid REDCap production per project line items that were sent
}
\examples{
\dontrun{sent_line_items <- get_unpaid_redcap_prod_per_project_line_items(rcc_billing_conn)}
}
