% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_rights.R
\name{get_user_rights_and_info}
\alias{get_user_rights_and_info}
\title{get_user_rights_and_info}
\usage{
get_user_rights_and_info(
  redcap_user_rights,
  redcap_user_roles,
  redcap_user_information
)
}
\arguments{
\item{redcap_user_rights, }{The contents of the REDCap table of the same name.}

\item{redcap_user_roles, }{The contents of the REDCap table of the same name.}

\item{redcap_user_information, }{The contents of the REDCap table of the same name.}
}
\value{
a dataframe of combined redcap_user_rights, roles, and user_information
}
\description{
Get redcap_user_rights combined with roles and user_information
}
\examples{
\dontrun{
get_user_rights_and_info(
  redcap_user_rights,
  redcap_user_roles,
  redcap_user_information
)
}
}
