% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invoice_line_item_communications_test_data}
\alias{invoice_line_item_communications_test_data}
\title{invoice_line_item_communications_test_data}
\format{
A data frame with 26 rows and 31 variables:
\describe{
  \item{\code{id}}{double: the primary key}
  \item{\code{service_identifier}}{character: either a redcap project ID, or redcap username}
  \item{\code{service_type_code}}{double: a numeric code that uniquely identifies the service_type}
  \item{\code{service_instance_id}}{character: a numeric code that uniquely identifies the service_instance}
  \item{\code{ctsi_study_id}}{double: CSBT's unique identifier for a service}
  \item{\code{name_of_service}}{character: name of the service}
  \item{\code{other_system_invoicing_comments}}{character: additional invoice information, either project url, or sponsor and pi}
  \item{\code{price_of_service}}{double: price of the service, in US dollars}
  \item{\code{qty_provided}}{double: quantity provided}
  \item{\code{amount_due}}{double: amount due, in US dollars}
  \item{\code{fiscal_year}}{character: fiscal year of the invoice}
  \item{\code{month_invoiced}}{character: month of the invoice}
  \item{\code{pi_last_name}}{character: last name of the person invoiced}
  \item{\code{pi_first_name}}{character: first name of the person invoiced}
  \item{\code{pi_email}}{character: email of the person invoiced}
  \item{\code{gatorlink}}{character: gatorlink of the person invoiced}
  \item{\code{crc_number}}{double: Clinical Research Center number}
  \item{\code{ids_number}}{character: Investigational Drug Service number}
  \item{\code{ocr_number}}{character: Office of Clinical Research study number}
  \item{\code{invoice_number}}{double: invoice number}
  \item{\code{je_number}}{character: journal entry number}
  \item{\code{je_posting_date}}{POSIXct: journal entry posting date}
  \item{\code{reason}}{character: reason for the invoice}
  \item{\code{status}}{character: status of the invoice}
  \item{\code{created}}{POSIXct: created at timestamp}
  \item{\code{updated}}{POSIXct: updated at timestamp}
  \item{\code{sender}}{character: message sender, typically an email address}
  \item{\code{recipient}}{character: message recipient, typically an email address}
  \item{\code{date_sent}}{POSIXct: date CTSI sent the message}
  \item{\code{date_received}}{POSIXct: date CTSIT received the email}
  \item{\code{script_name}}{character: the script that created this record}
}
}
\source{
\url{https://github.com/ctsit/rcc.billing/issues/7}
}
\usage{
invoice_line_item_communications_test_data
}
\description{
A test dataset for testing functions that write or read invoice_line_item_communications
}
\keyword{datasets}
