% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{invoice_line_item_df_from}
\alias{invoice_line_item_df_from}
\title{Creates a invoice_line_item data from invoice_line_item_communications_data}
\usage{
invoice_line_item_df_from(invoice_line_item_communications)
}
\arguments{
\item{invoice_line_item_communications, }{data that follows the format of invoice_line_item_communications_test_data located in R/data.R}
}
\value{
invoice_line_item dataframe
}
\description{
Creates a invoice_line_item data from invoice_line_item_communications_data
}
\examples{
\dontrun{
invoice_line_item_df_from(invoice_line_item_communications_test_data)
}
}
