% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invoice_line_item_reasons}
\alias{invoice_line_item_reasons}
\title{invoice_line_item_reasons}
\format{
A data frame with 3 rows and 2 variables:
\describe{
  \item{\code{code}}{character: A short, static name for the string}
  \item{\code{label}}{character: A friendly human-readable string}
}
}
\usage{
invoice_line_item_reasons
}
\description{
Provides text for new invoice line items
}
\keyword{datasets}
