% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invoice_line_item_statuses}
\alias{invoice_line_item_statuses}
\title{invoice_line_item_statuses}
\format{
A data frame with 5 rows and 2 variables:
\describe{
  \item{\code{status}}{character: A short, static name for the status code}
  \item{\code{description}}{character: A human-readable description of the status}
}
}
\usage{
invoice_line_item_statuses
}
\description{
Describes valid invoice line item statuses
}
\details{
DETAILS
}
\keyword{datasets}
