% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invoice_line_item_test_data}
\alias{invoice_line_item_test_data}
\title{Sample data for invoice_line_item table}
\format{
\describe{
  \item{\code{id}}{the primary key}
  \item{\code{service_identifier}}{either a redcap project ID, or redcap username}
  \item{\code{service_type_code}}{a numeric code that uniquely identifies the service_type}
  \item{\code{service_instance_id}}{a numeric code that uniquely identifies the service_instance}
  \item{\code{ctsi_study_id}}{CSBT's unique identifier for a service}
  \item{\code{name_of_service}}{name of the service}
  \item{\code{other_system_invoicing_comments}}{additional invoice information, either project url, or sponsor and pi}
  \item{\code{price_of_service}}{price of the service, in US dollars}
  \item{\code{qty_provided}}{quantity provided}
  \item{\code{amount_due}}{amount due, in US dollars}
  \item{\code{fiscal_year}}{fiscal year of the invoice}
  \item{\code{month_invoiced}}{month of the invoice}
  \item{\code{pi_last_name}}{last name of the person invoiced}
  \item{\code{pi_first_name}}{first name of the person invoiced}
  \item{\code{pi_email}}{email of the person invoiced}
  \item{\code{gatorlink}}{gatorlink of the person invoiced}
  \item{\code{reason}}{reason for the invoice}
  \item{\code{status}}{status of the invoice}
  \item{\code{sender}}{character: message sender, typically an email address}
  \item{\code{recipient}}{character: message recipient, typically an email address}
  \item{\code{date_sent}}{POSIXct: date CTSI sent the message}
  \item{\code{date_received}}{POSIXct: date CTSIT received the email}
  \item{\code{script_name}}{character: the script that created this record}
  \item{\code{created}}{created at timestamp}
  \item{\code{updated}}{updated at timestamp}
}
}
\source{
\url{https://github.com/ctsit/rcc.billing/issues/3}
}
\usage{
invoice_line_item_test_data
}
\description{
Sample data for invoice_line_item table
}
\keyword{datasets}
