% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uf_person_data.R
\name{is_faculty}
\alias{is_faculty}
\title{is_faculty}
\usage{
is_faculty(user_ids)
}
\arguments{
\item{user_ids, }{a vector of non-unique strings that are mostly UF Gatorlink IDs}
}
\value{
a logical vector the length of user_ids indicating which user_ids belong to faculty
}
\description{
A test to identify faculty in a vector of Gatorlink IDs
}
\examples{
\dontrun{
is_faculty(user_ids = c("pbc", "hoganwr", "cpb", "shapiroj", "pbc"))
[1] FALSE  TRUE FALSE  TRUE FALSE
}
}
