% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{mutate_columns_to_posixct}
\alias{mutate_columns_to_posixct}
\title{mutate_columns_to_posixct}
\usage{
mutate_columns_to_posixct(data, column_names)
}
\arguments{
\item{data}{- a dataframe to mutate}

\item{column_names}{- a vector of column names to mutate}
}
\value{
The input dataframe with revised data types
}
\description{
Mutates column data types to POSIXct
}
\examples{
\dontrun{
time_columns <- c("created", "updated")
mutate_columns_to_posixct(data, time_columns)
}
}
