% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{mysql_schema_to_sqlite}
\alias{mysql_schema_to_sqlite}
\title{converts an in-memory schema to a sqlite schema
and returns that schema.}
\usage{
mysql_schema_to_sqlite(schema)
}
\arguments{
\item{schema, }{a MySQL/MariaDB Schema}
}
\value{
sqlite schema for `schema`
}
\description{
converts an in-memory schema to a sqlite schema
and returns that schema.
}
\examples{
\dontrun{
mysql_schema_to_sqlite(schema)
}
}
