% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{next_n_months}
\alias{next_n_months}
\title{next_n_months}
\usage{
next_n_months(month, n = 1)
}
\arguments{
\item{month}{- an integer month number}

\item{n}{- the number of months to add to the current month (default = 1)}
}
\value{
the nth next month number as in integer
}
\description{
Return the month number that would occur n months after the integer in `month`
}
\examples{
next_n_months(9, 2)
next_n_months(1, 1)
}
