% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{previous_month}
\alias{previous_month}
\title{previous_month}
\usage{
previous_month(month)
}
\arguments{
\item{month}{- and integer month number}
}
\value{
the previous month number
}
\description{
Return the month number that would occur before the integer in `month`
}
\examples{
previous_month(9)
previous_month(1)
}
