% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{previous_n_months}
\alias{previous_n_months}
\title{previous_n_months}
\usage{
previous_n_months(month, n = 1)
}
\arguments{
\item{month}{- an integer month number}

\item{n}{- the number of months to subtract from current month (default = 1)}
}
\value{
the nth previous month number as in integer
}
\description{
Return the month number that would occur n months before the integer in `month`
}
\examples{
previous_n_months(9, 2)
previous_n_months(1, 1)
}
