% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{redcap_log_event_test_data}
\alias{redcap_log_event_test_data}
\title{redcap_log_event_test_data}
\format{
A named list of 10 dataframes each with a 0 or more rows of 15 variables.
  Each dataframe is structured as described here.
\describe{
  \item{\code{log_event_id}}{integer primary key}
  \item{\code{project_id}}{integer REDCap project_id}
  \item{\code{ts}}{double a timestamp in the format YYYYMMDDHHMMSS in local time}
  \item{\code{user}}{character the REDCap username who initiated this event}
  \item{\code{ip}}{character IP address of that REDCap user}
  \item{\code{page}}{character relative REDCap URL that initiated this event}
  \item{\code{event}}{character the type of event. e.g DATA_EXPORT, DELETE, ERROR, INSERT, MANAGE, OTHER, UPDATE}
  \item{\code{object_type}}{character a REDCap object type e.g., NA, redcap_arms, redcap_auth, redcap_data, redcap_docs, redcap_event_forms, redcap_events, redcap_events_metadata, redcap_metadata, redcap_metadata_temp, redcap_projects, redcap_user_rights}
  \item{\code{sql_log}}{character the SQL code executed during this event}
  \item{\code{pk}}{character the primary key of the effected object}
  \item{\code{event_id}}{integer the REDCap event_id if relevant}
  \item{\code{data_values}}{character}
  \item{\code{description}}{character description of the event}
  \item{\code{legacy}}{integer}
  \item{\code{change_reason}}{character}
}
}
\usage{
redcap_log_event_test_data
}
\description{
a named list of redcap_log_event tables with project lifecycle test data
}
\details{
DETAILS
}
\keyword{datasets}
