% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{redcap_rights_test_data}
\alias{redcap_rights_test_data}
\title{redcap_rights_test_data}
\format{
A named list of 3 dataframes:
\describe{
  \item{\code{redcap_user_information}}{REDCap Core table}
  \item{\code{redcap_user_rights}}{REDCap Core table}
  \item{\code{redcap_user_roles}}{REDCap Core table}
}
}
\usage{
redcap_rights_test_data
}
\description{
A named list of dataframes used to test the functions written for redcap_rights.R
}
\details{
DETAILS
}
\keyword{datasets}
