% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{redcap_user_information_test_data}
\alias{redcap_user_information_test_data}
\title{redcap_user_information_test_data}
\format{
A data frame with 5 rows and 49 variables:
\describe{
  \item{\code{ui_id}}{integer primary key}
  \item{\code{username}}{character redcap username}
  \item{\code{user_email}}{character primary email address}
  \item{\code{user_email2}}{character}
  \item{\code{user_email3}}{character}
  \item{\code{user_phone}}{character}
  \item{\code{user_phone_sms}}{character}
  \item{\code{user_firstname}}{character}
  \item{\code{user_lastname}}{character}
  \item{\code{user_inst_id}}{character}
  \item{\code{super_user}}{integer}
  \item{\code{account_manager}}{integer}
  \item{\code{access_system_config}}{integer}
  \item{\code{access_system_upgrade}}{integer}
  \item{\code{access_external_module_install}}{integer}
  \item{\code{admin_rights}}{integer}
  \item{\code{access_admin_dashboards}}{integer}
  \item{\code{user_creation}}{double}
  \item{\code{user_firstvisit}}{double}
  \item{\code{user_firstactivity}}{double}
  \item{\code{user_lastactivity}}{double}
  \item{\code{user_lastlogin}}{double}
  \item{\code{user_suspended_time}}{double}
  \item{\code{user_expiration}}{double}
  \item{\code{user_access_dashboard_view}}{double}
  \item{\code{user_access_dashboard_email_queued}}{character}
  \item{\code{user_sponsor}}{character}
  \item{\code{user_comments}}{character}
  \item{\code{allow_create_db}}{integer}
  \item{\code{email_verify_code}}{character}
  \item{\code{email2_verify_code}}{character}
  \item{\code{email3_verify_code}}{character}
  \item{\code{datetime_format}}{character}
  \item{\code{number_format_decimal}}{character}
  \item{\code{number_format_thousands_sep}}{character}
  \item{\code{csv_delimiter}}{character}
  \item{\code{two_factor_auth_secret}}{character}
  \item{\code{display_on_email_users}}{integer}
  \item{\code{two_factor_auth_twilio_prompt_phone}}{integer}
  \item{\code{two_factor_auth_code_expiration}}{integer}
  \item{\code{api_token}}{character}
  \item{\code{messaging_email_preference}}{character}
  \item{\code{messaging_email_urgent_all}}{integer}
  \item{\code{messaging_email_ts}}{double}
  \item{\code{messaging_email_general_system}}{integer}
  \item{\code{messaging_email_queue_time}}{double}
  \item{\code{ui_state}}{character}
  \item{\code{api_token_auto_request}}{integer}
  \item{\code{fhir_data_mart_create_project}}{integer}
}
}
\usage{
redcap_user_information_test_data
}
\description{
Test dataset for redcap_user_information
}
\details{
DETAILS
}
\keyword{datasets}
