% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sequester_project_test_data}
\alias{sequester_project_test_data}
\title{sequester_project_test_data}
\format{
A data frame with 11 rows and 3 variables:
\describe{
  \item{\code{name}}{character table name}
  \item{\code{schema}}{list of one table containing the schema for `name`}
  \item{\code{data}}{list of one tibble containing the data for `name`}
}
}
\usage{
sequester_project_test_data
}
\description{
A tibble of database tables used to test the functions written for project_sequestration.R
}
\details{
DETAILS
}
\keyword{datasets}
