% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_sequestration.R
\name{sequester_projects}
\alias{sequester_projects}
\title{sequester_projects}
\usage{
sequester_projects(conn, project_ids = as.numeric(NA))
}
\arguments{
\item{conn}{- a connection to a redcap database}

\item{project_ids}{- a vector of project IDs to be sequestered}
}
\value{
- a list describing the function activity via these objects
\itemize{
  \item project_ownership_sync_updates - updates made to project_ownership
  \item redcap_projects_sync_updates - updates made to redcap_projects
  \item project_ids_updated - project ids that received updates
}
}
\description{
sequester projects listed in `project_ids` that can be sequestered
}
\examples{
\dontrun{
sequester_projects(
  conn = rc_conn,
  project_ids = project_ids_to_sequester
}
}
