% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{service_instance_test_data}
\alias{service_instance_test_data}
\title{Sample data for service_instance table}
\format{
\describe{
  \item{service_instance_id}{the primary key}
  \item{service_type_code}{a numeric code that uniquely identifies the service_type}
  \item{service_identifier}{either a redcap project ID, or redcap username}
  \item{ctsi_study_id}{CSBT’s unique identifier for a service}
  \item{active}{a boolean indicating if we expect to continue billing for this service}
}
}
\source{
\url{https://github.com/ctsit/rcc.billing/issues/2}
}
\usage{
service_instance_test_data
}
\description{
Sample data for service_instance table
}
\keyword{datasets}
