% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{service_request_time}
\alias{service_request_time}
\title{Calculate Service Request Time}
\usage{
service_request_time(time_minutes, time_hours)
}
\arguments{
\item{time_minutes}{Numeric vector representing time in minutes.}

\item{time_hours}{Numeric vector representing time in hours.}
}
\value{
A numeric vector with the processed time in hours. For input times in minutes
        that are part of the set {15, 30, 45, 60}, the time is converted to hours.
        For time in hours greater than 1, the original hours are returned. Otherwise,
        `NA_real_` is returned for those cases not matching the conditions.
}
\description{
This function takes time in minutes and hours and returns a unified time format in hours.
}
\examples{
\dontrun{
service_request_time(30, 120)
}

}
