% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_ownership.R
\name{update_billable_by_ownership}
\alias{update_billable_by_ownership}
\title{Get a dataframe of updated billable status for project ownership projects,
set all projects as billable except those created by CTS-IT staff}
\usage{
update_billable_by_ownership(conn)
}
\arguments{
\item{conn}{- A REDCap database connection, e.g. the object returned from \code{\link[redcapcustodian]{connect_to_redcap_db}}}
}
\value{
A \code{\link{dataset_diff}} containing updates to project ownerhsip's "billable" column
}
\description{
Get a dataframe of updated billable status for project ownership projects,
set all projects as billable except those created by CTS-IT staff
}
\examples{
\dontrun{
conn <- redcapcustodian::connect_to_redcap_db()
project_ownership_changes <- update_billable_by_ownership(conn)
dbx::dbxUpdate(conn,
  table = "redcap_entity_project_ownership",
  records = updates$update_records,
  where_cols = c("id")
)
}
}
