% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_projects.R
\name{update_billable_if_owned_by_ctsit}
\alias{update_billable_if_owned_by_ctsit}
\title{Get a dataframe of updated billable status for project owned by CTS-IT staff}
\usage{
update_billable_if_owned_by_ctsit(conn)
}
\arguments{
\item{conn}{- A REDCap database connection, e.g. the object returned from \code{\link[redcapcustodian]{connect_to_redcap_db}}}
}
\value{
A \code{\link{dataset_diff}} containing updates to project ownerhsip's "billable" column
}
\description{
Ignore the current state billable and review every project owned by CTS-IT staff
}
\examples{
\dontrun{
conn <- redcapcustodian::connect_to_redcap_db()
billable_updates <- update_billable_if_owned_by_ctsit(conn)
dbx::dbxUpdate(conn,
  table = "redcap_entity_project_ownership",
  records = billable_updates$update_records,
  where_cols = c("id")
)
}
}
