% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{write_to_sqlite}
\alias{write_to_sqlite}
\title{Creates a table in sqlite for the provided table.
A corresponding data set and sqlite schema are required in /data and /schema respectively.
Use \code{\link{convert_schema_to_sqlite}} to generate a sqlite schema from a mysql schema.}
\usage{
write_to_sqlite(table_name)
}
\arguments{
\item{table_name, }{the name of the table to create in sqlite. A _test_data file in /data}
}
\description{
Creates a table in sqlite for the provided table.
A corresponding data set and sqlite schema are required in /data and /schema respectively.
Use \code{\link{convert_schema_to_sqlite}} to generate a sqlite schema from a mysql schema.
}
\examples{
\dontrun{
write_to_sqlite(table_name = "service_type")
}
}
