% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{convert_schema_to_sqlite}
\alias{convert_schema_to_sqlite}
\title{Locates a MySQL schema file for table_name, converts it to a sqlite schema
and returns that schema.}
\usage{
convert_schema_to_sqlite(table_name)
}
\arguments{
\item{table_name, }{the name of the table to convert}
}
\value{
sqlite schema for table_name
}
\description{
Locates a MySQL schema file for table_name, converts it to a sqlite schema
and returns that schema.
}
\examples{
\dontrun{
convert_schema_to_sqlite(table_name = "service_type")
}
}
