% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{create_table}
\alias{create_table}
\title{Creates a table based on a schema.}
\usage{
create_table(conn, schema)
}
\arguments{
\item{conn, }{a DBI connection object}

\item{schema, }{the ddl to execute against conn}
}
\description{
Creates a table based on a schema.
}
\examples{
\dontrun{
table_name <- "service_type"
conn <- DBI::dbConnect(RSQLite::SQLite(), dbname = ":memory:")

schema <- convert_schema_to_sqlite(table_name)
create_table(conn = conn, schema = schema)
}
}
