% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_scraping_utils.R
\name{get_bad_emails_from_log}
\alias{get_bad_emails_from_log}
\title{Get bad email addresses from the rcc.billing log}
\usage{
get_bad_emails_from_log(age_of_oldest_log_in_days = 8)
}
\arguments{
\item{age_of_oldest_log_in_days}{- an optional parameter indicating
age in days of the oldest log fie to be read. Defaults to 8 days.}
}
\value{
A vector of email addresses that resulted in an error in the rcc_job_log table
}
\description{
Get bad email addresses from the rcc.billing log
}
\examples{

\dontrun{
bad_recipients <- get_bad_emails_from_log(age_of_oldest_log_in_days = 31)
}
}
