% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_sequestration.R
\name{get_orphaned_projects}
\alias{get_orphaned_projects}
\title{get_orphaned_projects}
\usage{
get_orphaned_projects(conn, months_previous = 0)
}
\arguments{
\item{conn}{- a connection to a redcap database}

\item{months_previous}{- the nth month previous today to consider}
}
\value{
a dataframe describing orphaned projects
\itemize{
  \item project_id - project_id of the orphaned project
  \item reason - why this project was selected
  \item priority - the priority of the reason
}
}
\description{
Return a dataframe of projects that have been orphaned
}
\examples{
\dontrun{
get_orphaned_projects(
  conn = rc_conn,
  months_previous = 0
)
}
}
