% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probono_service_request_updates.R
\name{get_probono_service_request_updates}
\alias{get_probono_service_request_updates}
\title{Update Pro Bono Service Request Details}
\usage{
get_probono_service_request_updates(service_requests)
}
\arguments{
\item{service_requests}{A dataframe containing service request data from REDCap PID 1414}
}
\value{
Returns a dataframe with updated billable rate
}
\description{
This function processes a dataframe of service requests to update the billable_rate field
}
\examples{
\dontrun{
updates <- get_probono_service_request_updates(service_requests)
}
}
