% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_email.R
\name{get_processed_payment_data_from_email}
\alias{get_processed_payment_data_from_email}
\title{Scrape an inbox for processed payment emails and return data}
\usage{
get_processed_payment_data_from_email(
  username,
  password,
  url = "imaps://outlook.office365.com",
  messages_since_date
)
}
\arguments{
\item{username}{The username of the IMAP mailbox}

\item{password}{The password of the IMAP mailbox}

\item{url}{The IMAP URL of the host that houses the mailbox}

\item{messages_since_date}{The sent date of the oldest message that should be inspected}
}
\value{
A dataframe of processed payment facts
##\itemize{
##  \item email - stuff and things
##}
}
\description{
Connect to an imap mailbox, identify processed payment emails sent
after `messages_since_date`, and extract the data from those emails.
}
\examples{
\dontrun{
get_processed_payment_data_from_email(
  username = "jdoe",
  password = "jane_does_password",
  url ="imaps://outlook.office365.com",
  messages_since_date = as.Date("2022-01-01", format = "\%Y-\%m-\%d")
  )
}
}
