% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_details_for_billing.R
\name{get_project_details_for_billing}
\alias{get_project_details_for_billing}
\title{Get Project Details for Billing}
\usage{
get_project_details_for_billing(rc_conn, rc_billing_con, project_ids)
}
\arguments{
\item{rc_conn}{A REDCap database connection, e.g. the object returned from \code{\link[redcapcustodian]{connect_to_redcap_db}}}

\item{project_ids}{Vector of project IDs to retrieve details for.}

\item{rc_billing_conn}{A connection to REDCap billing database. \code{\link{connect_to_rcc_billing_db}}}
}
\value{
A data frame with project details.
}
\description{
This function retrieves detailed information about specific projects for billing purposes. It queries:
 - redcap_projects
 - redcap_entity_project_ownership
 - redcap_user_information
 - invoice_line_item
}
\examples{
\dontrun{
rc_conn <- connect_to_redcap_db()
rcc_billing_conn <- connect_to_rcc_billing_db()
project_ids <- c(12, 14, 22)
project_details <- get_project_details_for_billing(rc_conn, rc_billing_con, project_ids)
}

}
