% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_projects.R
\name{get_projects_needing_new_owners}
\alias{get_projects_needing_new_owners}
\title{get_projects_needing_new_owners}
\usage{
get_projects_needing_new_owners(
  redcap_entity_project_ownership,
  redcap_user_information
)
}
\arguments{
\item{redcap_entity_project_ownership, }{The contents of the REDCap Project Ownership table of the same name.}

\item{redcap_user_information, }{The contents of the REDCap table of the same name.}
}
\value{
a vector of project IDs
}
\description{
Returns the project IDs of projects that are owned by a REDCap user that has no primary email address
}
\examples{
\dontrun{
get_projects_needing_new_owners(
  redcap_entity_project_ownership = redcap_entity_project_ownership,
  redcap_user_information = redcap_user_information
)
}
}
