% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_projects.R
\name{get_projects_without_owners}
\alias{get_projects_without_owners}
\title{get_projects_without_owners}
\usage{
get_projects_without_owners(redcap_projects, redcap_entity_project_ownership)
}
\arguments{
\item{redcap_projects, }{The contents of the REDCap table of the same name.}

\item{redcap_entity_project_ownership, }{The contents of the REDCap Project Ownership table of the same name.}
}
\value{
a vector of project IDs
}
\description{
Returns the project_ids of projects without owners
}
\examples{
get_projects_without_owners(
  redcap_projects =
    cleanup_project_ownership_test_data$redcap_projects,
  redcap_entity_project_ownership =
    cleanup_project_ownership_test_data$redcap_entity_project_ownership
)
}
