% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_service_request_lines.R
\name{get_service_request_lines}
\alias{get_service_request_lines}
\title{Get Service Request Lines}
\usage{
get_service_request_lines(service_requests)
}
\arguments{
\item{service_requests}{A data frame of service requests, REDCap Service Request PID 1414.}
}
\value{
A data frame with response details to the service request
}
\description{
This function processes a dataset of service requests to extract and transform
various service details.
}
\examples{
\dontrun{
processed_requests <- get_service_request_lines(service_requests)
}
}
