% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_rights_and_info.R
\name{get_user_rights_and_info}
\alias{get_user_rights_and_info}
\title{Get every attribute of every permission entry and userinfo for each user on each permission}
\usage{
get_user_rights_and_info(
  rc_conn,
  require_active_account = T,
  require_active_permissions = T
)
}
\arguments{
\item{rc_conn}{A DBI connection object to a REDCap Database on a system.}

\item{require_active_account}{A boolean to indicate if only active accounts are returned. Defaults to TRUE.}

\item{require_active_permissions}{A boolean to indicate if only active permission entries are returned. Defaults to TRUE.}
}
\value{
a dataframe of every permission entry and with the matching user_information data appended
}
\description{
Get every attribute of every permission entry and userinfo for each user on each permission
}
\examples{
\dontrun{
get_user_rights_and_info(
  rc_conn = rc_conn,
  require_active_account = T,
  require_active_permissions = T
)
}
}
