% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequester_projects.R
\name{sequester_projects}
\alias{sequester_projects}
\title{sequester_projects}
\usage{
sequester_projects(
  conn,
  project_id = as.numeric(NA),
  reason = as.character(NA)
)
}
\arguments{
\item{conn}{- a connection to a redcap database}

\item{project_id}{- a vector of project IDs to be sequestered}

\item{reason}{- a vector of reasons the project IDs were sequestered}
}
\value{
- a list describing the function activity via these objects
\itemize{
  \item project_ownership_sync_updates - updates made to project_ownership
  \item redcap_projects_sync_updates - updates made to redcap_projects
  \item project_ids_updated - project ids that received updates
}
}
\description{
sequester projects listed in `project_ids` that can be sequestered
}
\examples{
\dontrun{
sequester_projects(
  conn = rc_conn,
  project_id = project_ids_to_sequester
  reason = reasons_project_ids_should_be_sequestered
}
}
