% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_invoice_line_items_for_csbt}
\alias{transform_invoice_line_items_for_csbt}
\title{Renames columns of a dataframe from CTSIT format to CSBT format}
\usage{
transform_invoice_line_items_for_csbt(invoice_line_items)
}
\arguments{
\item{invoice_line_items}{A dataframe with the CTSIT column names}
}
\value{
The input dataframe with columns adjusted to match CSBT format
}
\description{
Excludes non-CSBT columns and renames CTSIT column names to the corresponding CSBT names.
This function is the inverse of \code{\link{transform_invoice_line_items_for_ctsit}}
}
\details{
DETAILS
}
\examples{
\dontrun{
tbl(conn, "invoice_line_item") \%>\%
  collect() \%>\%
  transform_invoice_line_items_for_csbt()
}
}
\seealso{
\code{\link{csbt_column_names}}
}
