% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_invoice_line_items_for_ctsit}
\alias{transform_invoice_line_items_for_ctsit}
\title{Renames columns of a dataframe from CSBT format to CTSIT format}
\usage{
transform_invoice_line_items_for_ctsit(invoice_line_items)
}
\arguments{
\item{invoice_line_items}{A dataframe with the CSBT column names}
}
\value{
The input dataframe with columns adjusted to match CTSIT format
}
\description{
Renames CSBT column names to the corresponding CTSIT names.
This function is the inverse of \code{\link{transform_invoice_line_items_for_csbt}}, however it does NOT exclude columns not in CTSIT column names.
}
\details{
DETAILS
}
\examples{
\dontrun{
df_from_csbt \%>\%
  transform_invoice_line_items_for_ctsit() \%>\%
  janitor::clean_names()
}
}
\seealso{
\code{\link{csbt_column_names}}
}
